Added camera-loading-effect and improved model-loader
